controls/rovers/requirements.txt
coloredlogs==15.0.1
flatbuffers==20181003210633
humanfriendly==10.0
mpmath==1.3.0
numpy==2.3.1
onnxruntime==1.22.1
packaging==25.0
protobuf==6.31.1
RPi.GPIO==0.7.1
sympy==1.14.0